requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.1.0
cycler==0.11.0
datasets==2.12.0
diff-match-patch==20230430
dill==0.3.6
docker-pycreds==0.4.0
filelock==3.12.0
fire==0.5.0
flatbuffers==23.5.9
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
javac-parser==1.0.0
javalang==0.13.0
jax==0.4.10
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack-python==0.5.6
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
nltk==3.7
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.23.1
psutil==5.9.5
py4j==0.10.6
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.25.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stemming==1.0.1
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.3
Werkzeug==2.3.4
wrapt==1.14.1
xxhash==3.2.0
yarl==1.9.2
